Remove scalariform
Scalariform is now unmaintained and hasn’t been updated in a long time.
It starts to cause version conflicts with its dependency on old versions
of scala-xml, so I’m removing it.

I’ll set up scalafmt in another commit/PR, to keep the current diff and
change size smaller.