Run GHA builds on pull_request (not push)
Currently, every commit pushed to a pull request causes two builds,
which is unnecessary. Because using push can make pull request checks
misleading (where checks failed on an earlier commit but succeed on the
latest commit, even if the later commits have not fixed the problem that
caused the failure on the earlier commit), I’m limiting push to the main
branch.